fix check for mounted token (SOFTWARE-4371)
in theory the flock.opensciencegrid.org token can exist either in
/etc/condor/tokens-orig.d or in /etc/condor/tokens.d.  Perhaps the logic
changed at some point but I think the original idea was to copy all
tokens from tokens-orig.d to tokens.d and _then_ verify that the one
exists in tokens.d.  In any case i think this seems to do the trick.