Add a working syslog daemon to the container.
This (a) adds a syslog daemon to the container (via a setuid binary
as syslog is immovably at `/dev/log`) and (b) tells all HTCSS daemons
to log to the container.

The syslog daemon is set to log to the supervisord stdout, resulting
in all HTCondor logs getting to the container's stdout in a timely
manner.