use 'supervisorctl shutdown' instead of 'kill 1' (SOFTWARE-4608)
the supervisord does not run as pid 1 in the cvmfsexec case.

we can maybe get the pid from /var/run/supervisord.pid, if that's where
it lives (not clear if that's the case if running as non-root), but
apparently you can also just run the supervisorctl command, and that
case sounds the simplest to me.