Fix set_var() to not use the variable name as the value if the variable has an empty value
If a variable in the glidein config looked like
`OSG_SITE_NAME `
(i.e. with a trailing space but no value), then the awk expression in
set_var() would use `OSG_SITE_NAME ` as the value instead of the empty
string.

This was fixed in GlideinWMS somewhere between 3.9.6rc4 and 3.9.6 so we
should use their version once we can.