Change the default image OS from el8 (i.e., CentOS Stream 8) to al8 (i.e., Alma 8)
This changes the OS-less tags (such as 3.6-release) to be equivalent to the
corresponding al8 tags (so 3.6-al8-release).  It keeps the al8 tags for
backward compat, in case anybody is using those, it just adds the OS-less
tags as aliases.

Note that this has the side effect of making tags like 3.6-el8-release for
the CentOS Stream 8-based builds, which we didn't have before.