If POOL is a hostname or host:port then use it as a central manager as opposed to bailing out
We don't specify a separate CCB_ADDRESS in that case; also we disable remote syslog.
Also toss in a special case that if POOL is "ospool" then we choose prod-ospool or itb-ospool
depending on $ITB.  I like this better than having that behavior be implicit if POOL is blank.

(SOFTWARE-5383)