Do not try to capture syslog; this means we don't have to run rsyslog as root
When rsyslog was added (faf15501176d8f920c4afcac07464244df20ffe9),
we had configured Condor to log to syslog, and
then rsyslog needed to run as root in order to listen to the syslog
socket.

We switched to log tailing a while back
(46b3b07f18582a03ed6b2aa43cf5286e9baa1bbf) so this is no longer necessary.
In many cases (e.g. running via apptainer), it didn't work anyway.

To keep it simple, this commit does not remove the launch_rsyslogd
executable; we can remove it later.